requirements.txt
Django==4.1.3
python-dotenv==0.19.2
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
pymysql==1.1.0
dj-database-url==2.1.0
Pillow==10.0.1
django-redis==5.0.0